wca-git/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
PyGithub==2.5.0
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
